+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Release Versions:
 ## Upcoming changes
 
 - fix(modulo-controllers): improve log message (#110)
+- fix(modulo-components): add missing space in log message (#113)
 
 ## 4.2.1