+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Release Versions:
 
 - fix(modulo-controllers): improve log message (#110)
 - fix(modulo-components): add missing space in log message (#113)
+- feat(modulo-controller): add missing docs (#112)
 
 ## 4.2.1