+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Release Versions:
 - fix(modulo-controllers): improve log message (#110)
 - fix(modulo-components): add missing space in log message (#113)
 - feat(modulo-controller): add missing docs (#112)
+- refactor: rename function in modulo translators (#77)
 
 ## 4.2.1