+++ CHANGELOG.md
@@ -22,6 +22,7 @@ Release Versions:
 - refactor: rename function in modulo translators (#77)
 - feat: add mutex around step callback (#67)
 - refactor: move utilities to modulo_utils package (#89)
+- refactor: move exceptions to modulo_utils (#90)
 
 ## 4.2.1