+++ CHANGELOG.md
@@ -29,6 +29,7 @@ Release Versions:
 - feat(controllers): use exceptions from modulo utils and update to CL v8.1.0 (#106)
 - feat: define topic validation warning in utilities (#101)
 - fix(components): keep parameter values in sync with ros (#111)
+- feat: update to jazzy (#116)
 
 ## 4.2.1