+++ CHANGELOG.md
@@ -19,7 +19,7 @@ Release Versions:
 - fix(modulo-controllers): improve log message (#110)
 - fix(modulo-components): add missing space in log message (#113)
 - feat(modulo-controllers): add missing docs (#112)
-- fix(modulo-controllers): unhandled nullptr in read_input
+- fix(modulo-controllers): correctly handle nullptr in read_input (#118)
 
 ## 4.2.1